// src/types/LensProfileMatch.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DistortionCoeffs } from "./DistortionCoeffs";
import type { TcaCoeffs } from "./TcaCoeffs";
import type { VignettingCoeffs } from "./VignettingCoeffs";

export type LensProfileMatch = { profileId: string, lensName: string, distortion: DistortionCoeffs, tca: TcaCoeffs, vignetting: VignettingCoeffs | null, };